feat: tabbed simple and advanced modes
