+++ requirements.txt
@@ -6,4 +6,3 @@ pillow
 python-dotenv
 datasets
 huggingface_hub
-hashlib
